fix(APISIMP-SNAPPINNED-PAGECAP): lower SnapshotPinnedReadRest pageSize @Max(2000)→@Max(200), default 500→50 (XS) (#2477)
* fix(APISIMP-SNAPPINNED-PAGECAP): lower SnapshotPinnedReadRest pageSize @Max(2000)→@Max(200), default 500→50 (XS)

GET /v2/collections/{collectionAppId}/snapshots/{appId}/data-objects had
@Max(2000) with @DefaultValue("500") — both above the v2-standard cap of 200.
Lowering @Max without lowering the default would cause callers not passing
pageSize to implicitly send 500, failing Bean Validation with 400.

Changes:
- SnapshotPinnedReadRest.java: @Max(2000)→@Max(200), @DefaultValue("500")→@DefaultValue("50"),
  @Schema(maximum="2000",defaultValue="500")→@Schema(maximum="200",defaultValue="50"),
  @Operation and @Parameter descriptions updated
- SnapshotPinnedReadRestTest.java: new maxPageSize_isAtMost200 reflection test
  (asserts @Max ≤ 200 at the method-parameter level); paginates_beyondLastPage
  updated from pageSize=500/skip=5000 → pageSize=50/skip=500
- aidocs/16: APISIMP-SNAPPINNED-PAGECAP row added (in-flight)
- aidocs/34: APISIMP-UNHIDE-FEED-PAGECAP (merged fire-543 PR #2476) and
  APISIMP-SNAPPINNED-PAGECAP (in-flight) rows added
- aidocs/44: APISIMP-SNAPPINNED-PAGECAP row added

No /shepard/api/ v1 surface touched. No schema migration.


* fix(APISIMP-SNAPPINNED-PAGECAP): prevent CWE-190 int overflow in skip arithmetic

Use long skip = (long) page * pageSize in SnapshotPinnedReadRest to
prevent integer overflow at large page/pageSize combinations (CWE-190).
Update SnapshotService.listDataObjectAppIdsPage and SnapshotDAO.findDataObjectAppIds
to accept long skip; update test matchers to eq(0L/2L/500L) and anyLong().

---------

+++ backend/src/main/java/de/dlr/shepard/context/snapshot/daos/SnapshotDAO.java
@@ -331,7 +331,7 @@ public class SnapshotDAO extends GenericDAO<Snapshot> {
    * @param limit           maximum number of rows to return ({@code pageSize}).
    * @return paged list of DataObject appId strings; empty when skip &ge; total.
    */
-  public List<String> findDataObjectAppIds(long snapshotNeo4jId, int skip, int limit) {
+  public List<String> findDataObjectAppIds(long snapshotNeo4jId, long skip, int limit) {
     String query =
       "MATCH (e:SnapshotEntry)-[:ENTRY_OF]->(s:Snapshot) " +
       "WHERE id(s) = $id AND (e.deleted IS NULL OR e.deleted = false) " +
@@ -342,7 +342,7 @@ public class SnapshotDAO extends GenericDAO<Snapshot> {
       "SKIP $skip LIMIT $limit";
     org.neo4j.ogm.model.Result result = session.query(
       query,
-      Map.of("id", snapshotNeo4jId, "skip", (long) skip, "limit", (long) limit)
+      Map.of("id", snapshotNeo4jId, "skip", skip, "limit", (long) limit)
     );
     List<String> out = new ArrayList<>();
     for (Map<String, Object> row : result) {
